+++ client/package.json
@@ -31,9 +31,10 @@
         "dayjs": "^1.11.10",
         "emoji-mart": "^5.5.2",
         "express-rate-limit": "^7.1.5",
-        "faker": "^6.6.6",
+        "faker": "^5.5.3",
         "firebase": "^10.8.0",
         "firebase-admin": "^12.0.0",
+        "firebase-tools": "^13.2.1",
         "flowbite-react": "^0.7.0",
         "i18next": "^23.6.0",
         "i18next-browser-languagedetector": "^7.1.0",
@@ -63,7 +64,9 @@
         "@vitejs/plugin-react": "^4.0.3",
         "autoprefixer": "^10.4.16",
         "eslint": "^8.45.0",
+        "eslint-config-google": "^0.14.0",
         "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-import": "^2.29.1",
         "eslint-plugin-react-hooks": "^4.6.0",
         "eslint-plugin-react-refresh": "^0.4.3",
         "postcss": "^8.4.31",
