+++ client/src/components/.experiencia_chat/ConnectionManager.tsx
