+++ client/src/components/.experiencia_chat/MessagePanel.tsx
