+++ client/src/components/pintar_o_7/ProductThumbnail.tsx
@@ -9,7 +9,7 @@ import { checkLink } from '../../fetchers';
 export default function ProductThumbnail(props) {
     return (
         <Box component="div" sx={{ flexGrow: 1 }}>
-            <Link to={`/product/${props.product._id}`} state={props.product}>
+            <Link to={`/product/${props.product.id}`} state={props.product}>
                 <Box component="div" className="w-full aspect-square">
                     <img
                         className="w-full h-full aspect-square object-cover"
@@ -24,7 +24,7 @@ export default function ProductThumbnail(props) {
                     justifyContent: 'space-between',
                     width: '100%',
                 }}>
-                <Grid xs={8}>{props.product.title} </Grid>
+                <Grid xs={8}>{props.product.title}</Grid>
                 <Grid xs={4} sx={{ textAlign: 'right' }}>
                     {props.product.price}€
                 </Grid>
