+++ client/src/pages/NotFound.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Box, Divider, Typography } from '@mui/material';
+import { Box, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 
 const NotFound: React.FC = () => {
