+++ client/src/todo.md
@@ -1,0 +1,108 @@
+├── App.tsx [ ]
+├── components
+│   ├── CartBadge.tsx [ ]
+│   ├── CartDetails.tsx [ ]
+│   ├── CheckoutPaymentStep.tsx [ ]
+│   ├── CheckoutReviewStep.tsx [ ]
+│   ├── CheckoutShippingAddressStep.tsx [ ]
+│   ├── experinecia_chat [ ]
+│   │   ├── ChatForm.tsx [ ]
+│   │   ├── ChatMessage.tsx [ ]
+│   │   ├── ChatPage.tsx [ ]
+│   │   ├── Chat.tsx [ ]
+│   │   ├── ConnectionManager.tsx [ ]
+│   │   └── MessagePanel.tsx [ ]
+│   ├── Payment [ ]
+│   │   ├── MBWayComponent.tsx [ ]
+│   │   ├── MultibancoComponent.tsx [ ]
+│   │   └── PayPalComponent.tsx [ ]
+│   ├── pintar_o_7 [ ]
+│   │   ├── ArtistThumbnail.tsx [ ]
+│   │   ├── CountrySelect.tsx [ ]
+│   │   ├── Footer.tsx [ ]
+│   │   ├── Hero.tsx [ ]
+│   │   ├── NewProductRequest.tsx [ ]
+│   │   ├── Order.tsx [ ]
+│   │   ├── ProductThumbnail.tsx [ ]
+│   │   ├── ReactNavbar.tsx [ ]
+│   │   ├── SelectPrice.tsx [ ]
+│   │   └── SelectType.tsx [ ]
+│   ├── Product [ ]
+│   │   ├── ImageLightBox.tsx [ ]
+│   │   ├── ImageMagnifier.tsx [ ]
+│   │   ├── ProductAccordion.tsx [ ]
+│   │   └── ProductDetails.tsx [ ]
+│   ├── ProductCarousel [ ]
+│   │   ├── ProductCarousel.tsx [ ]
+│   │   └── styles.css [ ]
+│   ├── Profile [ ]
+│   │   ├── NotificationPaper.tsx [ ]
+│   │   └── ProfileThumbnail.tsx [ ]
+│   ├── ReservationDetails.tsx [ ]
+│   ├── ReserveModal.tsx [ ]
+│   └── Seller [ ]
+│   └── ProductPaper.tsx [ ]
+├── contexts [ ]
+│   ├── cartProvider.tsx [ ]
+│   ├── chatContext.tsx [ ]
+│   ├── currentAccountContext.tsx [ ]
+│   └── viewPortContext.tsx [ ]
+├── fetchers.tsx [ ]
+├── hooks [ ]
+│   ├── useArtistSearch.ts [ ]
+│   ├── useCart.tsx [ ]
+│   ├── useNotifications.ts [ ]
+│   ├── useProductSearch.ts [ ]
+│   └── useViewport.ts [ ]
+├── i18n.ts [ ]
+├── index.css [ ]
+├── locales [ ]
+│   ├── english.json [ ]
+│   └── portuguese.json [ ]
+├── main.tsx [x]
+├── pages [ ]
+│   ├── Administrator [ ]
+│   │   ├── Dashboard.tsx [ ]
+│   │   └── Requests.tsx [ ]
+│   ├── Cart.tsx [ ]
+│   ├── Checkout.tsx [ ]
+│   ├── Contact.tsx [ ]
+│   ├── InfoPage.tsx [ ]
+│   ├── NotFound.tsx [ ]
+│   ├── pintar_o_7 [ ]
+│   │   ├── ArtistsIndex.tsx [ ]
+│   │   ├── Artist.tsx [ ]
+│   │   ├── Home.tsx [ ]
+│   │   ├── Initial.tsx [ ]
+│   │   ├── Login.tsx [ ]
+│   │   └── Register.tsx [ ]
+│   ├── Product.tsx [ ]
+│   ├── Profile [ ]
+│   │   ├── Notifications.tsx [ ]
+│   │   ├── ProfileIndex.tsx [ ]
+│   │   ├── ProfileInfo.tsx [ ]
+│   │   └── ProfileOrderHistory.tsx [ ]
+│   └── Seller [ ]
+│   ├── NewProduct.tsx [ ]
+│   ├── NewSeller.tsx [ ]
+│   └── Products.tsx [ ]
+├── routes [ ]
+│   ├── AdminPrivateRoutes.tsx [ ]
+│   ├── PrivateRoutes.tsx [ ]
+│   └── SellerPrivateRoutes.tsx [ ]
+├── search [ ]
+│   └── search_products.ts [ ]
+├── socket.tsx [ ]
+├── types [ ]
+│   ├── country.ts [ ]
+│   ├── nestedPartial.ts [ ]
+│   ├── notification.ts [ ]
+│   ├── order.ts [ ]
+│   ├── product.ts [ ]
+│   ├── result.ts [ ]
+│   └── user.ts [ ]
+├── utils [ ]
+│   ├── db.js.bu [ ]
+│   ├── db.ts [ ]
+│   ├── firebase.js [ ]
+│   └── mock_data.js [ ]
