+++ client/src/types/product.ts
@@ -1,18 +1,18 @@
-
-export interface Product {
-    seller: string;
-    title: string;
-    author: string;
-    photos: string[];
-    description: string;
-    price: number;
-    product_type: string;
-    piece_info: PieceInfo | null;
-    // book_info: BookInfo | null;
-    published_date: Date;
-    //rating: number,     // [0-5]
-    //reviews: Review[],
-}
+export type Product = {
+  id: string;
+  seller: string;
+  title: string;
+  author: string;
+  photos: string[];
+  description: string;
+  price: number;
+  product_type: string;
+  piece_info: PieceInfo | null;
+  // book_info: BookInfo | null;
+  published_date: Date;
+  //rating: number,     // [0-5]
+  //reviews: Review[],
+};
 
 /*type Review = {
     id: number,
@@ -22,11 +22,11 @@ export interface Product {
 }*/
 
 export type PieceInfo = {
-    technique: string;
-    materials: string[];
-    dimensions: Dimensions;
-    year: number;
-    state: string;
+  technique: string;
+  materials: string[];
+  dimensions: Dimensions;
+  year: number;
+  state: string;
 };
 
 // type BookInfo = {
@@ -37,8 +37,14 @@ export type PieceInfo = {
 // };
 
 type Dimensions = {
-    height: number;
-    width: number;
-    depth: number;
-    weight: number;
+  height: number;
+  width: number;
+  depth: number;
+  weight: number;
 };
+
+export type ProductQuery = {
+  availabe?: boolean;
+  limit?: number;
+  featured?: boolean;
+}
